GEODE-1969: fix IOException on force flush
An oplog closed while writing to oplog with gemfire.syncWrites=true
whill no longer cause the force flush to fail with an
IOException.

The channel force call is now done inside the sync block
which makes sure that the write is flushed before another
thread can close the channel.

This closes #320